[libc][math][c23] Update input type.

+++ libc/src/math/fsqrtf16.h
@@ -14,7 +14,7 @@
 
 namespace LIBC_NAMESPACE_DECL {
 
-float16 fsqrtf16(float16 x);
+float16 fsqrtf16(float x);
 
 } // namespace LIBC_NAMESPACE_DECL
